Disable adding/removing additional IDs
For customers and in ZATCA business settings, we currently initialize the
additional ID fields and users are only expected to edit the values.

This is not good UX. We should probably have the ID type be a "Select"
lookup and allow users to add them as needed, but that will require some
validation work to ensure uniqueness and is not urgent enough for now.