Fix jinja error if taxes are not defined for any lines in the invoice
Previously, our tax logic only added the relevant fields if tax details were
found in the item wise tax details on taxes and charges template for the
invoice. That meant if this info was missing for any reason, our template would
raise an excpetion when trying to round the non-existent "total_amount" for the
item line.

We now properly set the tax percent and tax amount to 0 if they're missing.
This will generate a proper invoice which will later fail ZATCA validation
(e.g. during compliance)