Default to "Standard rate" if a sales invoice doesn't specify a tax category
We added a validation check to ensure a tax category is present, but that
doesn't work for old submitted invoices. Such invoices could have been
rejected for a variety of reasons, and we need to default to S tax category if
they don't specify one when attempting to fix the rejection.