Support fixing rejected sales invoices
A fixable rejection can happen in mainly two cases:

1) Bad or missing ZATCA configuration that leads to rejection. These
cases can be fixed by updating the configuration and generating another
'Sales Invoice Additional Fields' document for the invoice to submit to
ZATCA
2) An application bug that results in generating an invalid XML. These
cases can be fixed by updating the app to a later version that fixes the
issue, then generating another 'Sales Invoice Additional Fields' document
after the update to submit to ZATCA

This commit adds a custom button to the 'Sales Invoice Additional
Fields' document that allows users to trigger the aforementioned
generation.

As part of this change, the invoice counter was appended to the 'Sales
Invoice Additional Fields' name expression to ensure uniqueness.