Restore QR code in phase 2 standard invoice print format
It was removed assuming we're going to replace the QR code we generate
with one returned from the clearance API. On the sandbox, the clearance
API returns the same QR we generated in the XML, so there's no point in
doing this.

Even if clearance returns a new QR, we'll use it to fill the same field
in additional fields, so the print format should stay the same.