Fix parsing of ZATCA API responses
There were quite a few issues, at the heart of which is Python requests
overriding the truthiness of `Response`. We were checking `if response` in quite
a few places with the intention of detecting whether we have a response (i.e.
it's not None). Requests overrides truthiness to mean success (< 400 status
code), which resulted in quite a few code branches that were never hit before,
and thus had some problems.